+++ kataeventlib/src/main/java/software/daveturner/kataeventlib/PersonWriteEvent.java
@@ -1,0 +1,10 @@
+package software.daveturner.kataeventlib;
+
+public class PersonWriteEvent extends KataEvent{
+
+    public PersonWriteEvent(String body) {
+        super(body);
+        this.eventType = "PersonWriteEvent";
+        this.eventVersion = "1.0";
+    }
+}