+++ personwrite/src/main/java/software/daveturner/personwrite/api/V1ApiDelegateImpl.java
@@ -2,7 +2,6 @@ package software.daveturner.personwrite.api;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.context.request.NativeWebRequest;
 import software.daveturner.personwrite.model.Person;
 import software.daveturner.personwrite.service.PersonWriteService;
 
@@ -16,24 +15,6 @@ public class V1ApiDelegateImpl implements V1ApiDelegate{
     public V1ApiDelegateImpl(PersonWriteService service) {
         this.service = service;
     }
-
-    /*
-    @Override
-    public ResponseEntity<Person> createPerson(String name, Person person) {
-        return ResponseEntity.ok(savePerson(person));
-    }
-
-    @Override
-    public ResponseEntity<String> sayHello(String name) {
-        return ResponseEntity.ok("Hello " + name);
-    }
-
-    @Override
-    public ResponseEntity<Person> updatePerson(String name, Person person) {
-
-    }
-    */
-
     @Override
     public ResponseEntity<Void>  deletePerson(String id) {
         service.delete(id);
@@ -43,8 +24,7 @@ public class V1ApiDelegateImpl implements V1ApiDelegate{
     @Override
     public ResponseEntity<Person> fetchPerson(String personId) {
         Optional<Person> p = service.findById(personId);
-        if (p.isPresent()) return ResponseEntity.ok(p.get());
-        return ResponseEntity.notFound().build();
+        return p.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @Override