+++ personwrite/src/main/java/software/daveturner/personwrite/repo/PersonWriteRepo.java
@@ -2,7 +2,6 @@ package software.daveturner.personwrite.repo;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Component;
-import software.daveturner.personwrite.model.Person;
 import software.daveturner.personwrite.model.PersonWriteData;
 
 @Component