+++ personwrite/src/main/java/software/daveturner/personwrite/service/EventProducer.java
@@ -1,5 +1,7 @@
 package software.daveturner.personwrite.service;
 
+import software.daveturner.kataeventlib.KataEvent;
+
 public interface EventProducer {
-    void publishMessage(String jsonString);
+    void publishMessage(KataEvent event);
 }