+++ personwrite/src/main/java/software/daveturner/personwrite/service/PersonWriteService.java
@@ -1,13 +1,12 @@
 package software.daveturner.personwrite.service;
 
-import org.springframework.http.ResponseEntity;
 import software.daveturner.personwrite.model.Person;
 
 import java.util.Optional;
 
 public interface PersonWriteService {
-    public Person save(Person person);
-    public void delete(String id);
-    public Optional<Person> findById(String id);
-    public String fetchNewId();
+    Person save(Person person);
+    void delete(String id);
+    Optional<Person> findById(String id);
+    String fetchNewId();
 }