+++ personwrite/src/main/java/software/daveturner/personwrite/util/PersonWriteUtils.java
@@ -1,7 +1,12 @@
 package software.daveturner.personwrite.util;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.stereotype.Component;
-
+import software.daveturner.kataeventlib.KataEvent;
+import software.daveturner.kataeventlib.PersonDeleteEvent;
+import software.daveturner.kataeventlib.PersonWriteEvent;
+import software.daveturner.personwrite.model.Person;
 import java.util.Random;
 
 @Component
@@ -19,10 +24,26 @@ public class PersonWriteUtils {
 
     public String threeRandomNumbers() {
         Random r = new Random();
-        StringBuffer sb = new StringBuffer();
+        StringBuilder sb = new StringBuilder();
         sb.append(r.nextInt(10));
         sb.append(r.nextInt(10));
         sb.append(r.nextInt(10));
         return new String(sb);
+    }
+
+
+    public KataEvent createWriteEvent(Person person) {
+        ObjectMapper objectMapper = new ObjectMapper();
+        String jsonString;
+        try {
+            jsonString = objectMapper.writeValueAsString(person);
+            return new PersonWriteEvent(jsonString);
+        } catch (JsonProcessingException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public KataEvent createDeleteEvent(String id) {
+        return new PersonDeleteEvent(id);
     }
 }