+++ personwrite/src/test/java/software/daveturner/personwrite/util/PersonWriteUtilsTest.java
@@ -1,11 +1,8 @@
 package software.daveturner.personwrite.util;
 
-import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-
 import static org.apache.commons.lang3.StringUtils.*;
-import static org.junit.jupiter.api.Assertions.*;
 
 class PersonWriteUtilsTest {