+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 ## [0.0.2] - 2024-06-26
-- Mocking Kafka publish in PersonWrite tests
-- Initial Kafka Setup in PersonWrite
+- Kafka Support in PersonWrite
+- Mocking Kafka publish in PersonWrite tests and Cucumber
+
 
 ## [0.0.1] - 2024-06-26